Remove basic secrets provider
Since we do not expect much usage of this, we just nuke the basic file
and force use of encryption.
